[Stdlib] Unify Int with SIMD
This both changes the definition of `Int` over to an alias for
`SIMD[DType.int, 1]` and
migrates some usages of `Int` to a new `SIMDSize` type.

This new type is required for parameter inference to work in cases where
`SIMD` arguments
are passed with inferred width. It should otherwise not be commonly
used. SIMDSize is essentially the same as the "old" Int type, but may be
backed by a `pop.scalar<index>` in the future.

MODULAR_ORIG_COMMIT_REV_ID: cb60548c595da4ad798e9d17f50d7e516f90fc1a

+++ KGEN/test/mojo-integration/floats.mojo
@@ -49,8 +49,8 @@ def main():
     print(FloatLiteral.negative_zero)  # CHECK-NEXT: -0.0
     print(FloatLiteral.nan)  # CHECK-NEXT: nan
 
-    print(-1 * 0.0)  # CHECK-NEXT: -0.0
-    print(-1 * -0.0)  # CHECK-NEXT: 0.0
+    print(Float64(-1) * 0.0)  # CHECK-NEXT: -0.0
+    print(Float64(-1) * -0.0)  # CHECK-NEXT: 0.0
 
     # Very special cases of float literals were rounding incorrectly:
     # Specifically, where long division would end due to the mantissa being
